Implementado as repository e criado a primeira service --------
Retomada do projeto

Implementando Métodos
 - Sacar
 - Depositar
-----
Realizado ajuste no tipo da variavel conta alterado de double para BigDecimal
-----

+++ src/main/java/com/Bank/bank_system/Entity/Conta.java
@@ -6,6 +6,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Getter
 @Setter
@@ -21,7 +23,7 @@ public class Conta {
     @ManyToOne
     @JoinColumn(name = "cliente_id")
     private Cliente cliente;
-    private double saldo;
+    private BigDecimal saldo;
     private double limite;
     private String tipo;
     private String status;
